Add tokenizers
Tokenizers is the first part of a parsing library. Tokenizers are
responsible for breaking an unstructured input into meaniningful
tokens that can be later transformed to proper types.

This commit adds all tokenizers related to each line of the SDP
format. It also adds generalized tokenizers that can be used across
various lines, generalized by a const char which is the first letter
of the line. These generalized tokenizers will eventually replace
other non-generalized tokenizers, to minimize code complexity and
testing of different types that do pretty much the same thing.