Add SDP types
Types is the second step of a parsing libary. Types are concrete
structs or enums and represent the actual types that are defined
in the SDP RFC. Struct members naming follows that of the RFC.

Each type implements the `From<Tokenizer>` conversion that takes in
the associated type `Tokenizer` and generates the Type instance.
Transformations take place in that conversion as well, since
tokenizers hold only `&'a str` types. Hence, the conversion of any
integers, floats, or subtypes represented in other structs or enums
take place here.

Each type also implements the `Display` trait for the opposite of
parsing: generate a part of SDP message from the relevant type.

Each type might potentially implement some helper methods, for
instance the `Fmtp` type that represents the `fmtp` SDP attribute,
implements the `params_list` which tries to parse the
`format-specific-params` into an array, if they can be parsed like
that (depends on the format/params), returning a `Result`.

Note that this commit doesn't add any generalized parsing of an SDP
message. Instead, parsing (and `Display`ing) has been implemented on
the type level. The next step is to add a high level struct for
parsing an SDP message, that also holds the high level concepts of a
media description (which is different of the `Media` type that parses
just the `m=` line) etc.