Add high level types
While previous commit would add concrete types for each SDP line,
dealing with such types is very low level.

This commit adds high level types (and tokenizers) and most
importantly the `SessionDescription` type which represents the SDP
message as a whole. It also adds high level types for the `Time`
(which includes the active time, repeats and zone) and for the
`MediaDescription` which represents the media(s), in the concept
level (so not just the line).

`SessioDescription` type implements the `TryFrom` transformations from
the associated tokenizer, and also implements `Display` trait to
generate the SDP message. It also implements most common traits, with
most important being the `FromStr` trait for ease of use using the
`parse()` method when dealing with an `&str`.