Add better errors
Having meaningful errors is paramount when doing parsers. This commit
adds basic errors for better debugging. Specifically, whenever a
tokenizer or parser fails, the error will display the element that
the parser or tokenizer thought it was parsing, the actual input that
was parsed for that element and the underlying error that appeared
during parsing/tokenizing.